Return AdaCLIP features as float32
Return AdaCLIP features as float32

Cast prompted feature maps to float32 before passing them to the ResAD constraintor so CUDA autocast outputs do not conflict with float32 convolution parameters.

+++ models/adaclip_feature_extractor.py
@@ -261,7 +261,7 @@ class AdaCLIPPromptedFeatureExtractor(nn.Module):
             feature = features[key]
             if feature.dim() != 4:
                 raise ValueError(f"{key} must have shape [B,C,H,W], got {tuple(feature.shape)}")
-            ordered.append(feature)
+            ordered.append(feature.float())
         return ordered
 
     def _debug_shapes(self, features):
